期望页面：“预览”与之前的“详细”一样，只是名称变化一下。点击“修改按钮”后进入打分页面进行修改。修改完成后点击个人打分页面的“确定”按钮返回到下面页面进行提交。下面页面提交条件1.全部打完分可以提交；2.总分里相同分数的数量小于等于“2”个，可以提交。（预览与修改两个功能是否能融合一起，还是咱们弄，请陆哥帮忙）

+++ src/main/java/cn/ly/util/DateUtil.java
@@ -82,8 +82,8 @@ public class DateUtil {
 		int hr1 = (int) (ldate1 / 3600000); // 60*60*1000
 		int hr2 = (int) (ldate2 / 3600000);
 
-		int days1 = (int) hr1 / 24;
-		int days2 = (int) hr2 / 24;
+		int days1 = hr1 / 24;
+		int days2 = hr2 / 24;
 
 		long secsDiff = ldate2 - ldate1;
